Tweaked metas and moved script to the bottom
For http-equiv="content-type", I like to use that because it's a little more reassuring, though not absolutely necessary.
For viewport, I like it at the top because it's more of a browser-compatibility-management thing, as opposed to the actually informational metadata stuff below <title>.
For scripts, it's generally good practice (except in some cases like when you're loading your content via JS, which is terrible practice itself anyway) to load it at the bottom so that it doesn't block the rendering of your content. CSS needs to stay at the top because unstyled HTML is ugly af.